# Cache ts.program between linted files