## Background
This improves the performance issues outlined in #243.

## Approach
Rather than creating a watch program and having it parse each file as we lint it, instead we create a `ts.program` for each project and cache them.

When attempting to parse a file, we go through each created program and attempt to parse the given filename with it.

## Results
These are on my machine (2018 MBP w/ i7, 16GB RAM)
This repo w/o project
master: 3.40s
new: 3.24s

This repo w/ project
master: 13.34s
new: 5.68s

Single file in this repo w/ project
master: 10.83s
new: 4.92s

The repo originally mentioned in #243 w/ project
master: 50.72s
new: 9.94s

## Limitations
If a file that's being linted isn't included in the given tsconfig, performance remains poor, as we fall back to the current behaviour of parsing that file in isolation; https://github.com/eslint/eslint/issues/11222 might be able to help with this once it's shipped. This seems very much like an edge case though.