@JamesHenry 2019-03-19
Thanks so much for looking into this @dsgkirkby!

I was on site at Microsoft in Redmond today and paired with @uniqueiniquity on this area of the codebase.

We just merged an initial fix here: https://github.com/typescript-eslint/typescript-eslint/pull/367

There is still more work being done than in this PR, so it will be slower, but we would be grateful if you wouldn't mind installing the new canary version (`1.4.3-alpha.6`) and trying it on your private codebase for comparison?

Please could you also provide details on how you are benchmarking?

We can then assess where we're at and see the updated delta between your PR and the change we made today.

Thanks again!