@dsgkirkby 2019-03-20
@JamesHenry on `1.4.3-alpha6`, the run time on our codebase is now only about *10-11s*, which is a huge improvement, and only barely slower than this PR! 🎉 I'm actually quite shocked by just how fast it is for such a small change; I think I somewhat misidentified the cause of the slowness.

My benchmarking is simply running it normally on my laptop against the various repos (in the case of our private repo, via `yarn link`). The magnitude of performance difference meant that I didn't need to do anything too sophisticated.

The approach that you've taken is fundamentally different than what I've done here, as I completely removed the usage of watch mode and compiled normally; while the approach I took seems simpler overall (imo) it could also be somewhat risky to ship this as it's a larger change.

There are some small, utility-like parts of this PR that are probably worth keeping either way (unifying path resolution to a single codepath, choosing the first program that is able to parse the file rather than the first that's defined, the added tests), so one option if you'd like to keep the current approach would be for me to strip most of this PR and just leave that.