@uniqueiniquity 2019-04-03
Hi @dsgkirkby, sorry for taking so long to get back to you on this.
Thanks so much for validating the changes we made!

I do think we should keep the tests you added, since they're definitely adding coverage in a helpful way.
However, I'm not sure it's worth making the change to the usage of `firstDefined`. In particular, I think your change actually matches the current implementation - currently, it's taking the resulting list of programs from the first argument (i.e. `calculateProjectParserOptions) and returning the first program for which the result of the second argument (i.e. that anonymous function that calls `getSourceFile`) is defined. In my understanding, that's effectively exactly what your implementation does (i.e. computes the programs from the tsconfigs and then finds the first that had parsed that file), but let me know if I seem to be missing something. I think one thing to note there is that all the sourcefiles are parsed in `calculateProjectParserOptions` and not on the call to `getSourceFile`; that call simply just retrieves the already-parsed `SourceFile` object.

While I do appreciate the simplicity and clarity of your approach, I think we (maybe unfortunately) have to keep the existing approach. In particular, all this overhead is really just set up to manage two tricky pieces - `--fix` mode and non-standard file extensions (like `.vue`). The former is what primarily drove the usage of `WatchProgram`s; in particular, when using `--fix`, ESLint will lint a file multiple times, but not commit changes back to disk. This causes the `SourceFile` in the program to go out of sync with what it needs to actually be linting, since by default, that object is tied to what's on disk. By using `WatchProgram`s, we are effectively triggering our own updates to the program based on changes to the text of the file being linted that's stored in the memory of the ESLint process. 

The usage of the `WatchProgram` certainly seems to have some overhead as compared to purely generating our own programs. However, the benefit of having it handle those program updates (and the extra file extensions piece) seemed to outweigh the cost, rather than going ahead and reimplementing that ourselves.

Sorry for dumping that wall of text, but I hope that explains a little better why this seemingly heavy approach was needed. 

I really appreciate that you took the time to try to address the perf issue, and I think it would be great if we could still get those test changes in. Thanks again!