@uniqueiniquity 2019-04-04
@mistic Which version of typescript-estree are you using? There should be improvements in v1.5.0 or higher.