@mistic 2019-04-04
@uniqueiniquity according my `yarn.lock` file I'm using the very last version of everything related to `@typescript-eslint`

```
"@typescript-eslint/eslint-plugin@^1.6.0":
  version "1.6.0"
  resolved "https://registry.yarnpkg.com/@typescript-eslint/eslint-plugin/-/eslint-plugin-1.6.0.tgz#a5ff3128c692393fb16efa403ec7c8a5593dab0f"
  integrity sha512-U224c29E2lo861TQZs6GSmyC0OYeRNg6bE9UVIiFBxN2MlA0nq2dCrgIVyyRbC05UOcrgf2Wk/CF2gGOPQKUSQ==
  dependencies:
    "@typescript-eslint/parser" "1.6.0"
    "@typescript-eslint/typescript-estree" "1.6.0"
    requireindex "^1.2.0"
    tsutils "^3.7.0"

"@typescript-eslint/parser@1.6.0", "@typescript-eslint/parser@^1.6.0":
  version "1.6.0"
  resolved "https://registry.yarnpkg.com/@typescript-eslint/parser/-/parser-1.6.0.tgz#f01189c8b90848e3b8e45a6cdad27870529d1804"
  integrity sha512-VB9xmSbfafI+/kI4gUK3PfrkGmrJQfh0N4EScT1gZXSZyUxpsBirPL99EWZg9MmPG0pzq/gMtgkk7/rAHj4aQw==
  dependencies:
    "@typescript-eslint/typescript-estree" "1.6.0"
    eslint-scope "^4.0.0"
    eslint-visitor-keys "^1.0.0"

"@typescript-eslint/typescript-estree@1.6.0":
  version "1.6.0"
  resolved "https://registry.yarnpkg.com/@typescript-eslint/typescript-estree/-/typescript-estree-1.6.0.tgz#6cf43a07fee08b8eb52e4513b428c8cdc9751ef0"
  integrity sha512-A4CanUwfaG4oXobD5y7EXbsOHjCwn8tj1RDd820etpPAjH+Icjc2K9e/DQM1Hac5zH2BSy+u6bjvvF2wwREvYA==
  dependencies:
    lodash.unescape "4.0.1"
    semver "5.5.0"
```

I did some changes according to some other users' feedback also having that problem and as I was not using any typescript linting rule which required type information so  I just set the `parserOptions project key` to `undefined` and the liniting time recovered again from the same I was used to with tslint -> 6 minutes on CI. 

@uniqueiniquity do you think there would be any solution to solve that performance issue soon in the future in case we wanna use rules that actually requires type information and still doesn't increase too much the linting time?

I think one thing that might be helpful to reduce the number of affected users would be trying to understand by the given rules in a configuration if there is any of them that requires type information. If there is none just automatically set/override `parserOptions.project = undefined` or at least provide a warning advising to do so in the eslint typescript configuration. I think it would help a lot of people that is having that problem and that are not aware of the internals or didn't realise they doesn't have rules requiring types information and so they can disable that option to improve the linting performance.

Thanks for the quick answer on it @uniqueiniquity ! 👍